feat: LobbyService 생성
