+++ src/Game.ts
@@ -1,5 +1,6 @@
 import { ServerCommunicator } from "@tailhead/communicator";
 
+import LobbyService from "./LobbyService";
 import PlayerService from "./PlayerService";
 
 interface WordChecker {
@@ -10,6 +11,7 @@ export default class Game {
   constructor(
     private communicator: ServerCommunicator,
     private playerService: PlayerService,
+    private lobbyService: LobbyService,
     private wordChecker: WordChecker,
   ) {}
 
@@ -21,17 +23,9 @@ export default class Game {
     this.communicator.onReceive("exit", (playerId) => {
       this.playerService.leave(playerId);
     });
-  }
-
-  playerChat(playerId: string, content: string) {
-    const sendPlayer = this.playerService.getPlayer(playerId);
-    if (!sendPlayer) {
-      throw new Error("유령이다 유령");
-    }
-    const playersWithoutSender = this.playerService.getPlayers().filter((player) => player.id !== playerId);
-
-    playersWithoutSender.forEach((player) => {
-      this.communicator.sendOne(player.id, "playerChat", { playerId, nickname: sendPlayer.nickname, content });
+    this.communicator.onReceive("sendChat", (playerId, data) => {
+      this.lobbyService.playerChat(playerId, data.content);
     });
+    parseInt("", 10);
   }
 }
